requirements.txt
Django>=2.2
django-environ==0.8.1
django-extensions==3.1.5
django-test-plus==1.1.1
djangorestframework==3.12.4
djangorestframework-xml=2.0.0
factory_boy==2.11.1
Faker==0.9.2
flake8==3.5.0
flake8-bugbear==18.2.0
flake8-docstrings==1.3.0
ipython==7.16.2
pep8-naming==0.7.0
pre-commit==1.10.4
pytz==2018.5
